+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires =
     async-property>=0.2.1
     cryptography>=3.4.0
     httpcore<0.17.3
-    httpx[http2]==0.24.0
+    httpx[http2]>=0.19.0
     pydantic[dotenv]>=1.8.2,<3.0.0
     python-dateutil>=2.8.2
     readerwriterlock>=1.0.9
