+++ src/firebolt/async_db/connection.py
@@ -1,12 +1,9 @@
 from __future__ import annotations
 
-import socket
 from types import TracebackType
 from typing import Any, Dict, List, Optional, Type
 
-from httpcore.backends.auto import AutoBackend
-from httpcore.backends.base import AsyncNetworkStream
-from httpx import AsyncHTTPTransport, Timeout
+from httpx import Timeout
 
 from firebolt.async_db.cursor import Cursor, CursorV1, CursorV2
 from firebolt.async_db.util import _get_system_engine_url
@@ -14,11 +11,8 @@ from firebolt.client import DEFAULT_API_URL
 from firebolt.client.auth import Auth
 from firebolt.client.client import AsyncClient, AsyncClientV1, AsyncClientV2
 from firebolt.common.base_connection import BaseConnection
-from firebolt.common.settings import (
-    DEFAULT_TIMEOUT_SECONDS,
-    KEEPALIVE_FLAG,
-    KEEPIDLE_RATE,
-)
+from firebolt.common.http_backend import AsyncKeepaliveTransport
+from firebolt.common.settings import DEFAULT_TIMEOUT_SECONDS
 from firebolt.utils.exception import (
     ConfigurationError,
     ConnectionClosedError,
@@ -31,48 +25,6 @@ from firebolt.utils.util import (
     fix_url_schema,
     validate_engine_name_and_url_v1,
 )
-
-
-class OverriddenHttpBackend(AutoBackend):
-    """
-    `OverriddenHttpBackend` is a short-term solution for the TCP
-    connection idle timeout issue described in the following article:
-    https://docs.aws.amazon.com/elasticloadbalancing/latest/network/network-load-balancers.html#connection-idle-timeout
-    Since httpx creates a connection right before executing a request, the
-    backend must be overridden to set the socket to `KEEPALIVE`
-    and `KEEPIDLE` settings.
-    """
-
-    async def connect_tcp(  # type: ignore [override]
-        self,
-        host: str,
-        port: int,
-        timeout: Optional[float] = None,
-        local_address: Optional[str] = None,
-        **kwargs: Any,
-    ) -> AsyncNetworkStream:
-        stream = await super().connect_tcp(  # type: ignore [call-arg]
-            host,
-            port,
-            timeout=timeout,
-            local_address=local_address,
-            **kwargs,
-        )
-        # Enable keepalive
-        stream.get_extra_info("socket").setsockopt(
-            socket.SOL_SOCKET, socket.SO_KEEPALIVE, KEEPALIVE_FLAG
-        )
-        # MacOS does not have TCP_KEEPIDLE
-        if hasattr(socket, "TCP_KEEPIDLE"):
-            keepidle = socket.TCP_KEEPIDLE
-        else:
-            keepidle = 0x10  # TCP_KEEPALIVE on mac
-
-        # Set keepalive to 60 seconds
-        stream.get_extra_info("socket").setsockopt(
-            socket.IPPROTO_TCP, keepidle, KEEPIDLE_RATE
-        )
-        return stream
 
 
 class Connection(BaseConnection):
@@ -128,9 +80,6 @@ class Connection(BaseConnection):
         self.cursor_type = cursor_type
         self._cursors: List[Cursor] = []
         self._system_engine_connection = system_engine_connection
-        # Override tcp keepalive settings for connection
-        transport = AsyncHTTPTransport()
-        transport._pool._network_backend = OverriddenHttpBackend()
         self._client = client
 
     def cursor(self, **kwargs: Any) -> Cursor:
@@ -257,15 +206,13 @@ async def connect_v2(
         await _get_system_engine_url(auth, account_name, api_endpoint)
     )
 
-    transport = AsyncHTTPTransport()
-    transport._pool._network_backend = OverriddenHttpBackend()
     client = AsyncClientV2(
         auth=auth,
         account_name=account_name,
         base_url=system_engine_url,
         api_endpoint=api_endpoint,
         timeout=Timeout(DEFAULT_TIMEOUT_SECONDS, read=None),
-        transport=transport,
+        transport=AsyncKeepaliveTransport(),
         headers={"User-Agent": user_agent_header},
     )
     # Don't use context manager since this will be stored
@@ -312,7 +259,7 @@ async def connect_v2(
                 base_url=engine_url,
                 api_endpoint=api_endpoint,
                 timeout=Timeout(DEFAULT_TIMEOUT_SECONDS, read=None),
-                transport=transport,
+                transport=AsyncKeepaliveTransport(),
                 headers={"User-Agent": user_agent_header},
             )
             return Connection(
@@ -347,16 +294,13 @@ async def connect_v1(
 
     api_endpoint = fix_url_schema(api_endpoint)
 
-    # Override tcp keepalive settings for connection
-    transport = AsyncHTTPTransport()
-    transport._pool._network_backend = OverriddenHttpBackend()
     no_engine_client = AsyncClientV1(
         auth=auth,
         base_url=api_endpoint,
         account_name=account_name,
         api_endpoint=api_endpoint,
         timeout=Timeout(DEFAULT_TIMEOUT_SECONDS, read=None),
-        transport=transport,
+        transport=AsyncKeepaliveTransport(),
         headers={"User-Agent": user_agent_header},
     )
 
@@ -386,7 +330,7 @@ async def connect_v1(
         base_url=engine_url,
         api_endpoint=api_endpoint,
         timeout=Timeout(DEFAULT_TIMEOUT_SECONDS, read=None),
-        transport=transport,
+        transport=AsyncKeepaliveTransport(),
         headers={"User-Agent": user_agent_header},
     )
     return Connection(engine_url, database, client, CursorV1, None, api_endpoint)
